backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "django>=5.2.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "idna==3.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests==2.32.4",
    "urllib3==2.5.0",
]
